Document cargo-bdd release step
