+++ crates/rstest-bdd-macros/Cargo.toml
@@ -25,13 +25,14 @@ strict-compile-time-validation = ["compile-time-validation"] # Treats warnings a
 
 
 [dependencies]
+rstest-bdd-patterns.workspace = true
 proc-macro2 = { version = "1", features = ["span-locations"] }
 quote.workspace = true
 syn.workspace = true
 gherkin.workspace = true
 walkdir = "2.5.0"
 proc-macro-crate = "3"
-rstest-bdd.workspace = true
+regex.workspace = true
 cfg-if.workspace = true
 proc-macro-error = "1"
 
@@ -40,6 +41,7 @@ trybuild.workspace = true
 once_cell.workspace = true
 rstest.workspace = true
 serial_test.workspace = true
+rstest-bdd.workspace = true
 tempfile.workspace = true
 
 [package.metadata.docs.rs]
