+++ crates/rstest-bdd-patterns/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "rstest-bdd-patterns"
+version = "0.1.0-alpha3"
+edition.workspace = true
+license.workspace = true
+authors.workspace = true
+description = "Shared step-pattern compilation utilities for rstest-bdd."
+homepage.workspace = true
+repository.workspace = true
+readme = "README.md"
+keywords.workspace = true
+categories.workspace = true
+rust-version.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+regex.workspace = true
