+++ crates/rstest-bdd/Cargo.toml
@@ -16,6 +16,7 @@ rust-version.workspace = true
 workspace = true
 
 [dependencies]
+rstest-bdd-patterns.workspace = true
 gherkin.workspace = true
 inventory.workspace = true
 regex.workspace = true
